+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## Unreleased
+## 0.7.18 -- 28 Feb, 2026
 * `compose_files_into_zip_archive` now uses a temporary directory
+* Bugfix deprecated PIL's `font.getsize`
 
 ## 0.7.16 -- 22 Jan, 2026
 * Bugfix accidentally overriding python's `dict`